requirements.txt
streamlit
pandas
numpy
plotly
gspread
google-auth
google-auth-oauthlib
google-api-python-client